Update G09 Output Condenser
Finalized version to do both MOBCAL and ezSpectrum outputs that run (so long as the script is run on the same computer as ezSpectrum - I had trouble running this on our Linux server and then transferring the output to my laptop). 

Next update will attempt to implement a dictionary structure to remove some if statements and clean things up.